Separate states for boards, index page
