+++ src/components/SettingsPanel.vue
@@ -61,7 +61,8 @@
               label="Clear the board"
               no-caps
               text-color="white"
-              @click="clearBoard()"
+              :disable="!isBingoPage"
+              @click="clearBoard(currentRoute)"
             />
           </q-item-section>
         </q-item>
@@ -84,15 +85,26 @@
 </template>
 
 <script setup>
-
 // vue-related
-import { onMounted, watch } from 'vue'
+import { ref, onMounted, watch } from 'vue'
+import { useRoute, useRouter } from 'vue-router'
 
 // game settings
-import { gameSettings } from '../stores/gameSettings'
-import { gameState } from '../stores/gameState'
-const settings = gameSettings()
-const state = gameState()
+import { useGameStateStore } from '../stores/gameState'
+import { useGameSettingsStore } from '../stores/gameSettings'
+
+// route
+const route = useRoute()
+const router = useRouter()
+const currentRoute = ref(route.path.replace('/', ''))
+const isBingoPage = ref(currentRoute.value !== '')
+
+let state = null
+if (isBingoPage.value) {
+  state = useGameStateStore(currentRoute.value)
+}
+
+const settings = useGameSettingsStore()
 
 // font
 const updateFont = (removeFont) => {
@@ -104,25 +116,38 @@ const updateFont = (removeFont) => {
 }
 
 const clearBoard = () => {
-  if (confirm('Do you really want to clear the board state?')) {
+  if (confirm('Do you really want to clear this board state?')) {
     state.clearAll()
-    location.reload()
+    router.go()
   }
 }
 
 const clearAll = () => {
-  if (confirm('Do you really want to clear all the settings and board state?')) {
+  if (confirm('Do you really want to clear all the settings and all board states?')) {
     settings.clearAll()
-    state.clearAll()
-    location.reload()
+    localStorage.clear()
+    location.replace('/')
   }
 }
 
 onMounted(() => {
-  updateFont(settings.removeFont)
-})
+  watch(settings, (settings) => {
+    updateFont(settings.removeFont)
+  })
 
-watch(settings, (settings) => {
-  updateFont(settings.removeFont)
+  watch(
+    () => route.path, (newRoute) => {
+      console.log(newRoute)
+      currentRoute.value = route.path.replace('/', '')
+
+      if (currentRoute.value !== '') {
+        state = useGameStateStore(currentRoute.value)
+        isBingoPage.value = true
+      } else {
+        state = null
+        isBingoPage.value = false
+      }
+    }
+  )
 })
 </script>
