+++ src/css/app.scss
@@ -12,3 +12,38 @@ body {
 body:not(.remove-font) {
   font-family: 'FirstCoffee', "Roboto", "-apple-system", "Helvetica Neue", Helvetica, Arial, sans-serif;
 }
+
+.main {
+  margin: 1rem auto 0;
+  scroll-snap-align: center;
+}
+
+.index {
+  @extend .main;
+
+  width: 512px;
+}
+
+.bingo-page {
+  @extend .main;
+
+  width: 1000px;
+
+  body.screen--md & {
+    width: 85%;
+  }
+
+  body.screen--sm & {
+    width: 970px;
+  }
+}
+
+.fade-enter-active,
+.fade-leave-active {
+  transition: opacity 0.25s ease;
+}
+
+.fade-enter-from,
+.fade-leave-to {
+  opacity: 0;
+}
