+++ src/pages/ErrorNotFound.vue
@@ -1,5 +1,5 @@
 <template>
-  <div class="fullscreen bg-blue text-white text-center q-pa-md flex flex-center">
+  <div class="fullscreen bg-gymbag text-white text-center q-pa-md flex flex-center">
     <div>
       <div style="font-size: 30vh">
         404
@@ -9,13 +9,13 @@
         class="text-h2"
         style="opacity:.4"
       >
-        Oops. Nothing here...
+        No~
       </div>
 
       <q-btn
         class="q-mt-xl"
         color="white"
-        text-color="blue"
+        text-color="gymbag"
         unelevated
         to="/"
         label="Go Home"
